abstract card project
