+++ src/components/index.ts
@@ -1,2 +1,3 @@
 export { default as Layout } from './Layout'
 export { default as Sidebar } from './Sidebar'
+export { default as CardProject } from './CardProject'
